chore: moved rainEffect and thunderstromSoundEffect to MissionEnding
