+++ frontend/src/pages/InGame/components/MissionEnding.js
@@ -1,8 +1,32 @@
 import React, { useState, useEffect, useContext } from 'react';
 import { GameContext } from '../../../contexts';
+import { rainEffect } from '../Mission4/effect';
+
+import thunderstorm from '../../../assets/soundEffects/thunderstorm.mp3';
 import styled, { keyframes } from 'styled-components';
 
-const MissionEnding = ({ eventHandler, canvasRef }) => {
+const thunderstormSoundEffect = () => {
+  const volume = 0.5;
+  const audio = new Audio(thunderstorm);
+  audio.volume = volume;
+
+  // 사운드 재생
+  audio.play();
+
+  // 2초 후에 페이드 아웃 시작
+  setTimeout(() => {
+    const fadeOutInterval = setInterval(() => {
+      if (audio.volume <= 0.05) {
+        clearInterval(fadeOutInterval);
+        audio.pause(); // 오디오 재생 중지
+      } else {
+        audio.volume -= volume / 10; // 0.05
+      }
+    }, 100);
+  }, 2000);
+};
+
+const MissionEnding = ({ canvasRef }) => {
   const { isMissionEnding, setIsMissionEnding, myMissionStatus, inGameMode } =
     useContext(GameContext);
 
@@ -10,12 +34,10 @@ const MissionEnding = ({ eventHandler, canvasRef }) => {
     if (!isMissionEnding) return;
 
     console.log('💕💕💕IS MISSION ENDING MOUNTED!💕💕');
-    console.log('myMissionStatus:', myMissionStatus);
-    console.log('isMissionEnding:', isMissionEnding);
 
     if (inGameMode === 4 && !myMissionStatus) {
-      eventHandler.rainEffect(canvasRef, 2);
-      eventHandler.thunderstormSoundEffect();
+      rainEffect(canvasRef, 2);
+      thunderstormSoundEffect();
     }
   }, [isMissionEnding, myMissionStatus, setIsMissionEnding]);
